feat!: get tracklist from video files directly
This commit greatly changes the way we derive the tracklist. Here, we will probe
the video files directly using ffprobe and extract out the chapters, instead of
needing the user to prepare an XML/.txt file of chapters. As a consequence, this
module is codec agnostic and should work no matter what file it is as long as it
is a video file that can be ffprobe'd.

Irrelevant dependencies are also removed in this commit.