feat: add script checks for files
This commit adds pre-checks in the script for the number of arguments that are
fed to the script. If there are no arguments, i.e., no media files, then a
SyntaxError is raised to inform the user that they need to specify the file that
is to be probed. If there are more than 1 arguments, e.g., multiple media files,
then anoter SyntaxError is raised to inform the user that only one media file is
to be probed at a time.